Removed a few Linux incompatibilities. Added watchdog code to thread handler.
git-svn-id: svn://svn.code.sf.net/p/d2x-xl/code/trunk@817 01821946-851b-0410-b7f4-f1477c70e57a

+++ main/piggy.h
@@ -139,7 +139,7 @@ ubyte *LoadD1Palette (void);
 void UseBitmapCache (grsBitmap *bmP, int nSize);
 
 extern int bUseHiresTextures, bD1Data;
-extern unsigned int bitmapCacheUsed;
-extern unsigned int bitmapCacheSize;
+extern size_t bitmapCacheUsed;
+extern size_t bitmapCacheSize;
 
 #endif //_PIGGY_H
